+++ appyx-components/stable/spotlight/common/src/commonMain/kotlin/com/bumble/appyx/components/spotlight/ui/sliderscale/TargetUiState.kt
@@ -27,7 +27,7 @@ class TargetUiState(
     ) : this(
         positionInList = positionInList,
         position = Position.Target(
-            base.position.value.copy(
+            base.position.value.offset.copy(
                 x = (positionInList * elementWidth.value).dp
             )
         ),
@@ -47,9 +47,10 @@ class TargetUiState(
         return MutableUiState(
             uiContext = uiContext,
             position = Position(
-                uiContext, position,
-                scrollX.mapState(uiContext.coroutineScope) {
-                    DpOffset((it * elementWidth.value).dp, 0.dp)
+                uiContext = uiContext,
+                target = position,
+                displacement = scrollX.mapState(uiContext.coroutineScope) {
+                    Position.Value(offset = DpOffset((it * elementWidth.value).dp, 0.dp))
                 },
             ),
             scale = Scale(uiContext, scale,